Added wasm implementations for formattedtext
