+++ src/formattedtext/wasm_backend.rs
@@ -1,0 +1,12 @@
+use super::FormattedText;
+
+impl FormattedText {
+	pub fn newline() -> Result<(), ()> {
+		print!("\n");
+		return Ok(());
+	}
+
+	pub fn write(&self) -> String {
+		return self.text.clone();
+	}
+}
